Added the possibility to create users via POST.
This is still a very early prorotype, while learning golang, and obviously a http library would likely be used in the final version, and we would extract to several functions, have config for the DB connection etc. That is all for a later version. For now: YAY! IT WORKS \o/